refactor(web): extract misc hooks and utils from mid-size components
Continue the hook-extraction series for mid-size components:
- use-sign-in (shared): TanStack Form + authClient + toast out of
  sign-in-form.tsx
- use-add-player-search (live-sessions): useQuery + useDeferredValue +
  filter/reset state out of add-player-sheet.tsx
- use-assign-tournament (live-sessions): useMutation + useQuery + dialog
  state + create-and-assign flow out of assign-tournament-dialog.tsx
- session-events-formatters (live-sessions utils): pure formatters,
  payload summarizers, time bounds, and event grouping out of
  session-events-scene.tsx

use-assign-tournament uses invalidateTargets from optimistic-update.
Public props of every touched component are unchanged; callers
continue to work without modification.
